+++ extensions/simple-youdao/src/index.tsx
@@ -1,22 +1,21 @@
-import { useState, useRef, useCallback } from "react";
+import { useCallback, useRef, useState } from "react";
 import {
-  Icon,
-  Color,
-  LaunchProps,
-  List,
   Action,
   ActionPanel,
+  Color,
   getPreferenceValues,
-  Toast,
+  Icon,
+  LaunchProps,
+  List,
   showToast,
-  getSelectedText,
+  Toast,
 } from "@raycast/api";
-import { runAppleScript } from "run-applescript";
-import fetch, { Response, AbortError } from "node-fetch";
+import fetch, { AbortError, Response } from "node-fetch";
 import crypto, { randomUUID } from "crypto";
 import fs from "fs";
 import sound from "sound-play";
 import { URLSearchParams } from "url";
+import { exec } from "child_process";
 
 export default function Command(props: LaunchProps<{ arguments: Arguments.Index }>) {
   const { text } = props.arguments;
@@ -27,20 +26,62 @@ export default function Command(props: LaunchProps<{ arguments: Arguments.Index 
       searchText={searchText}
       isLoading={state.isLoading}
       onSearchTextChange={setSearchTextAndTranslate}
-      searchBarPlaceholder={
-        state.selection || state.clipboard
-          ? `default search from ${state.clipboard ? "Clipboard" : state.selection ? "Selection" : ""}${
-              state.searchText != "" ? " : " + state.searchText : ""
-            }`
-          : "input content wants to tranlate..."
-      }
-      throttle
+      searchBarPlaceholder="Enter text to translate"
     >
       {state.translateResult ? (
         <Translate translate_result={state.translateResult} state={state} setState={setState} />
       ) : null}
     </List>
   );
+}
+
+function parseError(code: number): string {
+  interface Message {
+    [key: number]: string;
+  }
+
+  const messages: Message = {
+    101: "缺少必填的参数",
+    102: "不支持的语言类型",
+    103: "翻译文本过长",
+    108: "应用ID无效",
+    110: "无相关服务的有效实例",
+    111: "开发者账号无效",
+    112: "请求服务无效",
+    113: "查询为空",
+    202: "签名检验失败,检查 KEY 和 SECRET",
+    401: "账户已经欠费",
+    411: "访问频率受限",
+  };
+  return messages[code] || "请参考错误码：" + code;
+}
+
+function isChinese(text: string | undefined): boolean {
+  if (!text) {
+    return false;
+  }
+  return /^[\u4e00-\u9fa5]+$/.test(text);
+}
+
+function parsePhonetic(basic: any, text: string | undefined): string {
+  if (!text) {
+    return "";
+  }
+  let phonetic = "";
+
+  if (isChinese(text) && basic.phonetic) {
+    phonetic = "[" + basic.phonetic + "] ";
+  }
+
+  if (basic["us-phonetic"]) {
+    phonetic += " [美: " + basic["us-phonetic"] + "] ";
+  }
+
+  if (basic["uk-phonetic"]) {
+    phonetic += " [英: " + basic["uk-phonetic"] + "]";
+  }
+
+  return phonetic;
 }
 
 function Translate({
@@ -53,13 +94,10 @@ function Translate({
   setState: React.Dispatch<React.SetStateAction<TranslateState>>;
 }) {
   if (translate_result && translate_result.errorCode && translate_result.errorCode !== "0") {
-    const errorMessage = `
-    * error code: ${translate_result.errorCode}.
-    * you can find all error code in here. (https://ai.youdao.com/DOCSIRMA/html/自然语言翻译/API文档/文本翻译服务/文本翻译服务-API文档.html)`;
     showToast({
       style: Toast.Style.Failure,
-      title: "Translation Error",
-      message: errorMessage,
+      title: "翻译出错啦",
+      message: parseError(Number(translate_result.errorCode)),
     });
   }
   return (
@@ -71,7 +109,9 @@ function Translate({
               key={index}
               title={item}
               subtitle={
-                translate_result.basic?.phonetic ? `${state.searchText} /${translate_result.basic?.phonetic}/` : ""
+                translate_result.basic?.phonetic
+                  ? `${state.searchText} ${parsePhonetic(translate_result.basic, state.searchText)}`
+                  : ""
               }
               icon={{ source: Icon.Dot, tintColor: Color.Red }}
               actions={
@@ -83,8 +123,7 @@ function Translate({
                   url={
                     translate_result.webdict && translate_result.webdict.url ? translate_result.webdict.url : undefined
                   }
-                  speak_url={translate_result.speakUrl}
-                  tspeak_url={translate_result.tSpeakUrl}
+                  pronounce={isChinese(state.searchText) ? item : state.searchText}
                 />
               }
             />
@@ -100,11 +139,14 @@ function Translate({
               icon={{ source: Icon.Dot, tintColor: Color.Blue }}
               actions={
                 <TranslateResultActionPanel
+                  setState={setState}
+                  text={state.searchText}
                   copy_content={item}
                   language={translate_result.l}
                   url={
                     translate_result.webdict && translate_result.webdict.url ? translate_result.webdict.url : undefined
                   }
+                  pronounce={isChinese(state.searchText) ? item : state.searchText}
                 />
               }
             />
@@ -121,11 +163,14 @@ function Translate({
               subtitle={item.key}
               actions={
                 <TranslateResultActionPanel
+                  setState={setState}
+                  text={state.searchText}
                   copy_content={item.value.join(", ")}
                   language={translate_result.l}
                   url={
                     translate_result.webdict && translate_result.webdict.url ? translate_result.webdict.url : undefined
                   }
+                  pronounce={isChinese(state.searchText) ? item.value[0] : item.key}
                 />
               }
             />
@@ -141,11 +186,10 @@ function TranslateResultActionPanel(props: {
   copy_content: string;
   language: string;
   url: string | undefined;
-  speak_url?: string;
-  tspeak_url?: string;
   setState?: React.Dispatch<React.SetStateAction<TranslateState>>;
+  pronounce?: string;
 }) {
-  const { text, copy_content, language, url, speak_url, tspeak_url, setState } = props;
+  const { text, copy_content, language, url, setState, pronounce } = props;
 
   //if need to use modern translation page
   const { is_using_modern_web } = getPreferenceValues();
@@ -158,18 +202,16 @@ function TranslateResultActionPanel(props: {
   return (
     <ActionPanel>
       <Action.CopyToClipboard content={copy_content} />
-      {webURL ? <Action.OpenInBrowser url={webURL} /> : null}
       <Action
         icon={Icon.Message}
         onAction={() => {
-          if (speak_url && setState) {
+          if (setState) {
             setState((oldState) => ({
               ...oldState,
               isLoading: true,
             }));
-            //try speak url first, and if it does not return 200, turn to use defaut service
             try {
-              pronunceIt(speak_url, text);
+              pronounceIt(pronounce);
             } catch (error) {
               console.log(error);
             } finally {
@@ -180,35 +222,10 @@ function TranslateResultActionPanel(props: {
             }
           }
         }}
-        shortcut={{ modifiers: ["ctrl"], key: "return" }}
-        title={"Read Original"}
+        shortcut={{ modifiers: ["cmd"], key: "return" }}
+        title={"Read"}
       />
-      {speak_url ? (
-        <Action
-          icon={Icon.Message}
-          onAction={() => {
-            if (speak_url && setState) {
-              setState((oldState) => ({
-                ...oldState,
-                isLoading: true,
-              }));
-              //try speak url first, and if it does not return 200, turn to use defaut service
-              try {
-                pronunceIt(tspeak_url, copy_content);
-              } catch (error) {
-                console.log(error);
-              } finally {
-                setState((oldState) => ({
-                  ...oldState,
-                  isLoading: false,
-                }));
-              }
-            }
-          }}
-          shortcut={{ modifiers: ["shift"], key: "return" }}
-          title={"Read Translated"}
-        />
-      ) : null}
+      {webURL ? <Action.OpenInBrowser shortcut={{ modifiers: ["ctrl"], key: "return" }} url={webURL} /> : null}
     </ActionPanel>
   );
 }
@@ -235,40 +252,18 @@ function useTranslate(argText: string) {
   const [state, setState] = useState<TranslateState>({
     searchText: argText,
     translateResult: undefined,
-    selection: false,
-    clipboard: false,
     isLoading: true,
   });
   const cancelRef = useRef<AbortController | null>(null);
 
   const translate = useCallback(
     async function translate(content: string) {
-      const { is_search_clipboard } = getPreferenceValues();
-      let isSelection = false;
-      let isClipboard = false;
-
-      try {
-        content = content || (await getSelectedText()).trim();
-        isSelection = !!content;
-      } catch (error) {
-        console.log("get selected text error...");
-      }
-
-      try {
-        if (!content && is_search_clipboard) {
-          content = await runAppleScript("the clipboard");
-          isClipboard = !!content;
-        }
-      } catch (error) {
-        console.log("get clipboard text error...");
-      }
-
-      if (content && content.length < 50) {
+      const { max_input_length } = getPreferenceValues();
+      const maxContentLength = /^[1-9]\d*$/.test(max_input_length) ? Number(max_input_length) : 500;
+      if (content && content.length < maxContentLength) {
         content = content.trim();
       } else {
         content = "";
-        isSelection = false;
-        isClipboard = false;
       }
 
       cancelRef.current?.abort();
@@ -276,8 +271,6 @@ function useTranslate(argText: string) {
       setState((oldState) => ({
         ...oldState,
         isLoading: true,
-        selection: isSelection,
-        clipboard: isClipboard,
         searchText: content,
       }));
       try {
@@ -325,8 +318,7 @@ async function performTranslate(searchText: string, signal: AbortSignal): Promis
 function generateSign(content: string, salt: string, curtime: number, app_key: string, app_secret: string) {
   const sha256 = crypto.createHash("sha256");
   sha256.update(app_key + getContentForSign(content) + salt + curtime + app_secret);
-  const cipher = sha256.digest("hex");
-  return cipher;
+  return sha256.digest("hex");
 }
 
 function getContentForSign(content: string) {
@@ -340,7 +332,6 @@ function translateAPI(content: string, signal: AbortSignal): Promise<Response> {
   const q = content;
   const salt = randomUUID();
   const curtime = Math.floor(Date.now() / 1000);
-  content;
   const sign = generateSign(q, salt, curtime, app_key, app_secret);
   const query = new URLSearchParams([
     ["q", q],
@@ -361,30 +352,38 @@ function translateAPI(content: string, signal: AbortSignal): Promise<Response> {
   });
 }
 
-function pronunceIt(speak_url: string | undefined, speak_text: string | undefined): void {
-  if (speak_url == undefined || speak_url.length === 0) {
+function sayIt(speak_text: string) {
+  return new Promise<boolean>((resolve) => {
+    exec(`say ${speak_text} -v Samantha`, (error) => {
+      if (error) {
+        console.log("failed to say it", error);
+        resolve(false);
+      } else {
+        resolve(true);
+      }
+    });
+  });
+}
+
+function pronounceIt(speak_text: string | undefined): void {
+  if (speak_text == undefined || speak_text.length === 0) {
     return;
   }
-  fetch(speak_url)
-    .then((res) => {
-      if (res.status !== 200 && res.headers.get("Content-Type") === "audio/mp3" && speak_text != undefined) {
-        return fetch(`http://dict.youdao.com/dictvoice?audio=${speak_text}`);
-      } else {
-        return res;
-      }
-    })
-    .then((res) => {
-      const fileStream = fs.createWriteStream("/tmp/tmp_raycast_simpleyd.mp3");
-      res?.body?.pipe(fileStream);
-      sound.play("/tmp/tmp_raycast_simpleyd.mp3");
-    });
+
+  sayIt(speak_text).then((success) => {
+    if (!success) {
+      fetch(`http://dict.youdao.com/dictvoice?type=0&audio=${speak_text}`).then((res) => {
+        const fileStream = fs.createWriteStream("/tmp/tmp_raycast_simpleyd.mp3");
+        res?.body?.pipe(fileStream);
+        sound.play("/tmp/tmp_raycast_simpleyd.mp3");
+      });
+    }
+  });
 }
 
 interface TranslateState {
   translateResult?: TranslateResult;
   searchText?: string;
-  clipboard?: boolean;
-  selection?: boolean;
   isLoading: boolean;
 }